Refactor routes and enhance markdown handling
Reorganized route structure and improved markdown rendering. Added new routes for tech, idea, and career sections along with breadcrumb navigation. Integrated markdown-to-HTML conversion for better content display, updated dependencies, and renamed several files for consistency.
