+++ app/routes/_app.tech/route.tsx
@@ -1,12 +1,9 @@
-import { Link } from '@remix-run/react';
-import { View } from '~/routes/_app._index/view';
-import { useAppLoaderData } from '~/routes/_app/useAppLoaderData';
+import { Link, Outlet } from '@remix-run/react';
 
 export const handle = {
   Breadcrumb: () => <Link to={"/tech"}>tech</Link>,
 };
 
 export default function Route() {
-  const { list } = useAppLoaderData();
-  return <View list={list.filter((item) => item.path.startsWith("tech"))} />;
+  return <Outlet />;
 }
