+++ app/routes/_app.tech_.$sha/route.tsx
@@ -1,5 +1,6 @@
-import { Link, useLoaderData } from "@remix-run/react";
-import { loaderFunction } from "~/routes/_app._index._$sha/loader.server";
+import { Link, useLoaderData } from '@remix-run/react';
+import { loaderFunction } from '~/routes/_app._index._$sha/loader.server';
+import { View } from '~/routes/_app._index._$sha/view';
 
 export const handle = {
   Breadcrumb: ({ sha }: Awaited<ReturnType<typeof loader>>) => <Link to={`/tech/${sha}`}>{sha}</Link>,
@@ -8,6 +9,6 @@ export const handle = {
 export const loader = loaderFunction;
 
 export default function Route() {
-  const { markdown } = useLoaderData<typeof loader>();
-  return <pre>{markdown}</pre>;
+  const { html } = useLoaderData<typeof loader>();
+  return <View html={html} />;
 }
