+++ package.json
@@ -31,7 +31,9 @@
     "react-dom": "^18.2.0",
     "tailwind-merge": "^2.5.5",
     "tailwindcss-animate": "^1.0.7",
-    "tiny-invariant": "^1.3.3"
+    "tiny-invariant": "^1.3.3",
+    "zenn-content-css": "^0.1.158",
+    "zenn-markdown-html": "^0.1.158"
   },
   "devDependencies": {
     "@biomejs/biome": "1.9.4",
